Use u16 instead of u32 for pixel coords
About 3-5% decrease in performance, but the API is now semantically
correct as TGA image dimensions are stored in u16s.

+++ tinytga/src/pixel.rs
@@ -2,10 +2,10 @@
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Default)]
 pub struct Pixel {
     /// Pixel X coordinate from top left of image
-    pub x: u32,
+    pub x: u16,
 
     /// Pixel Y coordinate from top left of image
-    pub y: u32,
+    pub y: u16,
 
     /// Pixel color
     pub color: u32,
